DellReplicate.jl

Documentation for DellReplicate. HELLO IS THIS WORKING ?

DellReplicate.figure1_data_cleaner

DellReplicate.figure1_visualise

DellReplicate.gen_vars_fig1!

DellReplicate.read_csv

DellReplicate.figure1_data_cleaner — Method

figure1datacleaner() Loads the climate_panel_csv dataset and reproduces Dell's (2012) makefigure1.do commands. Returns a DataFrame object which can be used by ???

DellReplicate.figure1_visualise — Method

figure1visualise(df::String) Plots Figure 1 from Dell (2012) by calling the data cleaning function `figure1datacleanerwith theclimatepanel_csv.csv` dataset.

DellReplicate.gen_vars_fig1! — Method

gen_vars_fig1!(df::DataFrame)

Generates the necessary mean temperature and precipiation variables for the two graphs of Figure 1, given the climate panel data. Returns the modified version of input df.

DellReplicate.read_csv — Method

read_csv(fn::String)

Creates a DataFrame object from a .csv file, where fn is the file name. May only work if ran from a directory where assets if is in the same parent directory.