Functions for Figure 1.

This page contains the functions used to generate Figure 1 of Dell (2012).

DellReplicate.figure1_data_cleaner

DellReplicate.figure1_visualise

DellReplicate.gen_vars_fig1!

DellReplicate.make_table1

DellReplicate.read_csv

DellReplicate.figure1_visualise — Function

figure1_visualise(df::String)

Plots Figure 1 from Dell (2012) by calling the data cleaning function figure1_data_cleaner with the climate_panel_csv.csv dataset. The figure is a combination of 128 line plots (one for each country) showing the temperature range and two scatter plots showing the mean temperature values for the periods 1950-1959 and 1996-2005.

DellReplicate.figure1_data_cleaner — Function

figure1_data_cleaner()

Loads the climate_panel_csv dataset and reproduces Dell's (2012) makefigure1.do commands. Returns a DataFrame object which can be used by figure1_visualise().

DellReplicate.read_csv — Function

read_csv(fn::String)

Creates a DataFrame object from a .csv file, where fn is the file name. May only work if ran from a directory where assets if is in the same parent directory.

DellReplicate.gen_vars_fig1! — Function

gen_vars_fig1!(df::DataFrame)

Generates the necessary mean temperature and precipiation variables for the two graphs of Figure 1, given the climate panel data. Returns the modified version of input df.