Commonly used functions in DellReplicate.jl

This page contains functions which are used in more than one figure/table.

DellReplicate.read_csv — Function

read_csv(fn::String)

Creates a DataFrame object from a .csv file, where fn is the file name. May only work if ran from a directory where assets if is in the same parent directory.

DellReplicate.growth_var! — Function

growth_var!(df::DataFrames.DataFrame, g_var_name::Symbol, var_name::Symbol)

Adds a new column to df with the name g_var_name computing the growth of var_name.

Missing docstring.

Missing docstring for keep_20years_gdp. Check Documenter's build log for details.

DellReplicate.filter_transform! — Function

filter_transform!(df::DataFrames.DataFrame, pred, args)

Transforms the data on a filtered subset using a predicate function (a function that returns true or false).

DellReplicate.gen_lag_vars — Function

gen_lag_vars(df::DataFrames.DataFrame)

Generates all the lagged variables as well as the interaction between temperature/precipitation and poor/rich variables necessary for figure 2 and others.

DellReplicate.gen_year_vars — Function

gen_year_vars(df::DataFrames.DataFrame)

This function follows uses loops extensively to generate interaction variables between region and year, as well as dummy variables for each year in the dataset.

DellReplicate.gen_xtile_vars — Function

gen_xtile_vars(climate_panel::DataFrames.DataFrame)

This function creates copies of the DataFrame and dummy variables classifying each country as poor or rich as well as less/more hot. Returns a merged DataFrame with the new columns.