Functions for Table 1.

This page contains the specific functions used to create Table 1. Some functions which are common to multiple tables/figures may not be present.

Table 1

Statistic 	Quarter 	Half 	ThreeQuarter 	One 	One_and_quarter 	One_and_half

String 	Float64 	Float64 	Float64 	Float64 	Float64 	Float64

Raw Data 	0.573 	0.299 	0.144 	0.064 	0.028 	0.011

Without year FE 	0.511 	0.215 	0.085 	0.032 	0.013 	0.005

Table 2

Statistic 	One 	Two 	Three 	Four 	Five 	Six

String 	Float64 	Float64 	Float64 	Float64 	Float64 	Float64

Raw Data 	0.48 	0.229 	0.121 	0.07 	0.042 	0.027

Without year FE 	0.494 	0.221 	0.113 	0.062 	0.038 	0.024

DellReplicate.make_table1 — Function

make_table1(raw_df_name::String)

This function transform the data and produce summary statistics. It returns two DataFrame and presents them as pretty tables. The first table, table1, shows the proportion of country with temperature above or below country mean with a certain bin, and the second table, table2, is giving the same information about the precipation with 100mm units for the thresholds.