Functions for Table 2.

The main function for Table 2 is make_table2 which calls various functions a shown below.

DellReplicate.make_table2 — Function

make_table2(raw_df_name::String)

The "master" function used to store OLS coefficients and standard errors from Table 2 in the paper. Performs various data cleaning actions and relies on the same functions as make_table1() to generate the variables.

DellReplicate.qr_method — Function

qr_method(X::Matrix; columns::Dict)

Returns a new matrix of regressors free of collinearity issues (i.e. full rank matrix) using the QR method as described by Engler (1997). Also returns a dictionnary with the name of the regressor as a key and its new position in the regressors matrix as a value, from the dict columns which maps an old position to a regressor name. Note that the new_correspondance dict can also be used to retrieve standard errors (done later in the code).

DellReplicate.create_cluster — Function

create_cluster(Y::Vector, cluster_values::Vector)

Generates a size(Y)xsize(Y) matrix of indicator variables equal to 1 if observation i and j share the same cluster value. The cluster_value parameter accepts vectors of the same size as Y.

DellReplicate.two_way_clustered_sterrs — Function

two_way_clustered_sterrs(cluster::Matrix, X::Matrix, Y::Vector, β::Vector)

This function computes the variance-covariance matrix of regressors contained in X, with their associated coefficients β. The formula comes from Cameron et al. (2011). In particular, we use equations (2.5) and (2.8) and create the "cluster matrix", a NxN (where N is the number of obs. in each specification) indicator matrix with ijth entry equal to one if observations i and j share any cluster (country and region-year), and 0 if not.

DellReplicate.check_coeffs_table2 — Function

check_coeffs_table2(df_julia::DataFrames.DataFrame, df_stata::DataFrames.DataFrame; other_regressors)

Given a DataFrame and the base specification of Table 2, this function computes the OLS coefficients and associated standard errors. Accepts other regressors (i.e. those in all columns but column 1), which must be passed as an array. It returns a dict with two keys: Julia and Stata. Suppose there are K regressors per specification. Each key maps to a dict containing K other dicts, each one associated with a regressor containing two key-value pairs: the OLS coefficient and the standard error.

Examples

ols_coeff_wtempoorcountry = col2["julia"]["wtem_initxtilegdp1"]["coeff"]
-1.6551452111665383

sterr_wtempoorcountry = col2["stata"]["wtem_initxtilegdp1"]["st. error"]
0.45663109132426655

Missing docstring.

Missing docstring for figure1_visualise_graph2. Check Documenter's build log for details.