Functions for Table 2.

The main function for Table 2 is make_table2 which calls various functions as shown below.

Dependent variable: annual GDP growth rate 	Model 1 	Model 2 	Model 3 	Model 4 	Model 5

Temperature 	-0.325414 	0.260944 	0.262417 	0.171924 	0.563024

(0.2667381063964909) 	(0.292172378732773) 	(0.29127925123569204) 	(0.27461688629574804) 	(0.29776243180916123)

Poor country dummy 	 	-1.65515 	-1.60954 	-1.64475 	-1.79081

(0.4566307366223894) 	(0.4554073052767715) 	(0.45389741776568265) 	(0.43387893311917186)

Hot country dummy 	 	 	 	0.2369

(0.5351042496329304)

Agricultural country dummy 	 	 	 	 	-0.379312

(0.3895778806537779)

Precipitation 	 	 	-0.0830812 	-0.228171 	-0.0984184

(0.04264245661412808) 	(0.062045675612721986) 	(0.04449462158369184)

Poor country dummy 	 	 	0.152895 	0.160498 	0.153419

(0.06711216733230993) 	(0.06287833488955635) 	(0.07336304702910652)

Hot country dummy 	 	 	 	0.184694

(0.06567424550564989)

Agricultural country dummy 	 	 	 	 	-0.00803258

(0.06841961679884542)

Standard errors are in parentheses. They differ slighlty from those reported in the paper due to the way Stata computes them. The method we chose does not allow us to apply a correction to the computation of B in the function two_way_clustered_sterrs(). In particual, following the method proposed by Cameron et al. (2011), we would have to multiply $u$ by $\sqrt{c}$ where $c = \frac{G}{G-1}$ and $G$ is equal to the number of observations in each cluster. However, the number of observations in the interaction of the clusters country and region-year is often 1, resulting in an error. This produces a downward bias on our computed standard errors, which remain fairly close to the results reported by the authors.

DellReplicate.make_table2 — Function

make_table2(raw_df_name::String)

The "master" function used to store OLS coefficients and standard errors from Table 2 in the paper. Performs various data cleaning actions and relies on the same functions as make_table1() to generate the variables. Finally, it constructs a PrettyTables.pretty_table object which prints the coeffs and their standard errors in the terminal matching the style of Table 2 in the paper.

DellReplicate.qr_method — Function

qr_method(X::Matrix; columns::Dict)

Returns a new matrix of regressors free of collinearity issues (i.e. full rank matrix) using the QR method as described by Engler (1997). Also returns a dictionnary with the name of the regressor as a key and its new position in the regressors matrix as a value, from the dict columns which maps an old position to a regressor name. Note that the new_correspondance dict can also be used to retrieve standard errors (done later in the code).

DellReplicate.create_cluster — Function

create_cluster(Y::Vector, cluster_values::Vector)

Generates a size(Y)xsize(Y) matrix of indicator variables equal to 1 if observation i and j share the same cluster value. The cluster_value parameter accepts vectors of the same size as Y.

DellReplicate.two_way_clustered_sterrs — Function

two_way_clustered_sterrs(cluster::Matrix, X::Matrix, Y::Vector, β::Vector)

This function computes the variance-covariance matrix of regressors contained in X, with their associated coefficients β. The formula comes from Cameron et al. (2011). In particular, we use equations (2.5) and (2.8) and create the "cluster matrix", a NxN (where N is the number of obs. in each specification) indicator matrix with ijth entry equal to one if observations i and j share any cluster (country and region-year), and 0 if not.

DellReplicate.check_coeffs_table2 — Function

check_coeffs_table2(df_julia::DataFrames.DataFrame, df_stata::DataFrames.DataFrame; other_regressors)

Given a DataFrame and the base specification of Table 2, this function computes the OLS coefficients and associated standard errors. Accepts other regressors (i.e. those in all columns but column 1), which must be passed as an array. It returns a dict with two keys: Julia and Stata. Suppose there are K regressors per specification. Each key maps to a dict containing K other dicts, each one associated with a regressor containing two key-value pairs: the OLS coefficient and the standard error.

Examples

ols_coeff_wtempoorcountry = col2["julia"]["wtem_initxtilegdp1"]["coeff"]
-1.6551452111665383

sterr_wtempoorcountry = col2["stata"]["wtem_initxtilegdp1"]["st. error"]
0.45663109132426655